requirements.txt
PySide6
qfluentwidgets
qframelesswindow
python-docx
openpyxl
# 安装轻量版
pip install PySide6-Fluent-Widgets -i https://pypi.org/simple/

# 安装完整版
pip install "PySide6-Fluent-Widgets[full]" -i https://pypi.org/simple/
